feat(excel-checker): spreadsheet value parser
